refactor: reorganize code to work with rye and maturin
