+++ python/codecrafters_sqlite/varint.py
