--- test_main.py
@@ -1,7 +1,0 @@
-import doctest
-
-import app
-
-
-def test_docstring():
-    assert doctest.testmod(m=app).failed == 0
